Fix 5734: Set alarm clock to switch profile (from ics - NEW)
ics: http://review.cyanogenmod.com/#/c/20271/2

This change creates a new setting for the alarm, that allows to associate
a profile with an alarm, and change to this profile when the alarm triggers.

Update change for reflect new profiles status (enabled/disabled)

NOTE: The database of alarms is upgrading from version 6 to version 7, so all
alarms are resetting with this change.

Change-Id: I3d20f0e406b23c1d4c8d873434c331b3425971e4